+++ packages/blocks/core/CHANGELOG.md
@@ -77,6 +77,7 @@
 - `Blocksy → Contact Info` block: inner blocks improved.
 - `Blocksy → Dynamic Data` block: inner blocks improved.
 - `Blocksy → Search` block: inner blocks improved.
+- `Archive` block: support improved.
 
 
 ### Automated Tests
@@ -91,6 +92,7 @@
 - Updated E2E tests to check functionality of `Query Pagination Next` block.
 - Updated E2E tests to check functionality of `Query Pagination Numbers` block.
 - Updated E2E tests to check functionality of `Query Pagination Prev` block.
+- Updated E2E tests to check support of `Archive` block.
 
 
 ## 1.3.0 (2025-04-27)